ReplicaSets: Better handling of server item's refresh and replica set folder item's refresh for set's unreachable cases. Adding code to show more correct error string for replica set connections. Updating script engine's dbname dynamically.

+++ src/robomongo/core/settings/ConnectionSettings.h
@@ -72,7 +72,7 @@ namespace Robomongo
         /**
          * @brief Port of server
          */
-        unsigned short serverPort() const { return _port; }
+        int serverPort() const { return _port; }
         void setServerPort(const int port) { _port = port; }
 
         /**
